Add all available gems
I propose this change, to have a faster update cycle. Very often we release a new version of e.g. _deliver_, but don't immediately push a new _fastlane_ release out, so that way the user would get the update available message, and we tell them to run `fastlane update_fastlane`, which would only update _fastlane_ and not _deliver_. 
By adding all tools to this list, the update would be installed for the user 👍