Restore fastlane command shims
<rant>
Pull request #60 (Deprecate the tool shims and ask users to execute `fastlane <tool>`) did not *deprecate* the tool shims, it just broke thousands of scripts relying on running the fastlane tools directly.

What’s the point of providing a shim just to tell the user that they should run the command differently?

I think that every time we (programmers) tell the user what they should do in order to resolve an issue instead of doing it automatically, we have failed. Sometimes we ask the user to do something because doing it programmatically would be a lot of work. For fastlane tools, doing the right thing, the thing that the user expects is literally one line!

```
exec "`dirname "${BASH_SOURCE[0]}"`/bundle/bin/bundle-env" fastlane <tool> "$@"
```

Last year, I started as a .NET developer after 8 years doing iOS development. I had to learn a lot of things, so naturally I read a lot of blog posts and Stack Overflow answers. The most frustrating experience was reading something that I knew would solve my problem but didn’t because the tools or API changed. Please don’t do that with fastlane.
</rant>